lazy-load eas lib
