+++ components/common/DatabaseEditor/components/table/table.spec.tsx
@@ -20,8 +20,6 @@ beforeEach(() => {
   FetchMock.fn.mockReset();
 });
 
-jest.mock('@ethereum-attestation-service/eas-sdk', () => ({}));
-
 jest.mock('next/router', () => ({
   useRouter: () => ({
     pathname: '/[domain]/',
