+++ lib/credentials/getAttestation.ts
@@ -1,11 +1,9 @@
-import { JsonRpcProvider } from '@ethersproject/providers';
 import { getChainById } from '@root/connectors/chains';
-import { optimism } from 'viem/chains';
 
 import { getEthersProvider } from '../blockchain/getEthersProvider';
 
-import { getEasInstance, type EasSchemaChain } from './connectors';
-import { decodeOptimismProjectSnapshotAttestation } from './schemas/optimismProjectSchemas';
+import { type EasSchemaChain } from './connectors';
+import { getEasInstance } from './getEasInstance';
 
 export async function getAttestation({ attestationUID, chainId }: { chainId: EasSchemaChain; attestationUID: string }) {
   const eas = getEasInstance(chainId);
