+++ lib/credentials/getEasInstance.ts
@@ -1,0 +1,8 @@
+import { EAS, getSchemaUID } from '@ethereum-attestation-service/eas-sdk';
+
+import type { EasSchemaChain } from './connectors';
+import { getEasConnector } from './connectors';
+
+export function getEasInstance(chainId: EasSchemaChain) {
+  return new EAS(getEasConnector(chainId).attestationContract);
+}
