+++ lib/credentials/schemas/reward.ts
@@ -14,27 +14,3 @@ export type RewardCredential = {
   rewardURL: string;
   Event: string;
 };
-
-export function encodeRewardCredential({ Description, Name, Organization, Event, rewardURL }: RewardCredential) {
-  const encoder = new SchemaEncoder(rewardCredentialSchemaDefinition);
-  const encodedData = encoder.encodeData([
-    { name: 'Name', value: Name, type: 'string' },
-    { name: 'Organization', value: Organization, type: 'string' },
-    { name: 'Description', value: Description, type: 'string' },
-    { name: 'rewardURL', value: rewardURL, type: 'string' },
-    { name: 'Event', value: Event, type: 'string' }
-  ] as TypedSchemaItem<RewardCredential>[]);
-
-  return encodedData;
-}
-
-export function decodeRewardCredential(rawData: string): RewardCredential {
-  const decoder = new SchemaEncoder(rewardCredentialSchemaDefinition);
-  const parsed = decoder.decodeData(rawData);
-  const values = parsed.reduce((acc, item) => {
-    acc[item.name as keyof RewardCredential] = item.value.value as string;
-    return acc;
-  }, {} as RewardCredential);
-
-  return values as RewardCredential;
-}
